Detection_Code
This code is used to send pings from an ultra-sonic sensor and test/organize data all received. The Arduino reads and tests the data received from the ping to determines the DOT based on the time and distance measurements. A temporary sample structure is created for each object. If the object is in the same zone within a period of time, then data will be collected into a sample structure. The number of data samples and the overall number of sample collections on each object is counted. 
Next steps will be to process the collected data.