SD Card Read & Write
Code enables user to define the rate at which data will be displayed by using a COMMANDS.txt file that will be created on the SD Card. Arduino can then write the desired data to the SD Card at the user's commanded-rate. The text document containing the data will be saved as LOG.cav on the SD Card.

-Tutorial 11: SD Cards and Data Logging by Jeremy Blum was used to create this code.